refactor(login): rename component to include sing-in & sing-up text field attributes
